# qualify-backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/analysts": {
            "get": {
                "description": "Retorna uma lista de analistas com filtros opcionais",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Analistas"
                ],
                "summary": "Listar analistas",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Nome parcial para busca",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "País",
                        "name": "country",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Cidade",
                        "name": "city",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "Valor mínimo por hora",
                        "name": "min_hourly_rate",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "Valor máximo por hora",
                        "name": "max_hourly_rate",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Quantidade mínima de avaliações totais",
                        "name": "min_total_reviews",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "Avaliação média mínima",
                        "name": "min_mean_rating",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Campo para ordenar: name,country_name,city,hourly_rate,total_reviews,mean_rating,time_created",
                        "name": "sort_by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Direção: ASC ou DESC",
                        "name": "order",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pkg.AnalystsResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/users/{id}/analyst": {
            "get": {
                "description": "Retorna os detalhes de um analista pelo ID do usuário",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Analistas"
                ],
                "summary": "Obter analista",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID do usuário",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pkg.AnalystResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "Atualiza dados do analista pelo ID do usuário",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Analistas"
                ],
                "summary": "Atualizar analista",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID do usuário",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Objeto analista",
                        "name": "analyst",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pkg.Analyst"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pkg.AnalystResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Atribui o papel de analista a um usuário existente",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Analistas"
                ],
                "summary": "Criar papel de analista",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID do usuário",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Objeto analista (envie apenas `hourly_rate`)",
                        "name": "analyst",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pkg.Analyst"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/pkg.AnalystResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "Remove o papel de analista de um usuário",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Analistas"
                ],
                "summary": "Excluir analista",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID do usuário",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "pkg.Analyst": {
            "type": "object",
            "properties": {
                "city": {
                    "type": "string"
                },
                "country_code": {
                    "type": "string"
                },
                "country_name": {
                    "type": "string"
                },
                "country_state": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "hourly_rate": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "mean_rating": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "time_created": {
                    "type": "string"
                },
                "timezone": {
                    "type": "string"
                },
                "total_reviews": {
                    "type": "integer"
                }
            }
        },
...
        "pkg.AnalystResponse": {
            "type": "object",
            "properties": {
                "analyst": {
                    "$ref": "#/definitions/pkg.Analyst"
                }
            }
        },
...
        "pkg.AnalystsResponse": {
            "type": "object",
            "properties": {
                "analysts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/pkg.Analyst"
                    }
                },
                "count": {
                    "type": "integer"
                }
            }
        },
...
    }
}